+++ app/src/main/AndroidManifest.xml
@@ -23,7 +23,7 @@
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".ChatListActivity"
+            android:name=".activities.ChatListActivity"
             android:label="@string/app_name">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -32,24 +32,24 @@
         </activity>
 
         <activity
-            android:name=".ChatDetailActivity"
+            android:name=".activities.ChatDetailActivity"
             android:label="@string/title_chat_detail"
-            android:parentActivityName=".ChatListActivity">
+            android:parentActivityName=".activities.ChatListActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ChatListActivity" />
+                android:value=".activities.ChatListActivity" />
         </activity>
 
-        <activity android:name=".LoginActivity" />
+        <activity android:name=".activities.LoginActivity" />
 
         <service android:name="com.nbusy.organizer.data.DataService" />
 
-        <service android:name=".GcmIntentService" />
+        <service android:name=".services.GcmIntentService" />
 
-        <service android:name=".WorkerService" />
+        <service android:name=".worker.WorkerService" />
 
         <receiver
-            android:name=".GcmBroadcastReceiver"
+            android:name=".services.GcmBroadcastReceiver"
             android:permission="com.google.android.c2dm.permission.SEND">
             <intent-filter>
                 <action android:name="com.google.android.c2dm.intent.RECEIVE" />
@@ -57,7 +57,7 @@
             </intent-filter>
         </receiver>
 
-        <receiver android:name=".DeviceBootBroadcastReceiver">
+        <receiver android:name=".services.DeviceBootBroadcastReceiver">
             <intent-filter>
                 <action android:name="android.intent.action.BOOT_COMPLETED" />
             </intent-filter>
