+++ app/src/main/java/com/nbusy/app/activities/LoginActivity.java
@@ -1,4 +1,4 @@
-package com.nbusy.app;
+package com.nbusy.app.activities;
 
 import android.app.Activity;
 import android.content.Intent;
@@ -13,6 +13,7 @@ import com.google.android.gms.common.Scopes;
 import com.google.android.gms.common.api.GoogleApiClient;
 import com.google.android.gms.common.api.Scope;
 import com.google.android.gms.plus.Plus;
+import com.nbusy.app.R;
 
 public class LoginActivity extends Activity implements
         GoogleApiClient.ConnectionCallbacks,
