+++ app/src/main/java/com/nbusy/app/data/Chat.java
@@ -1,4 +1,6 @@
-package com.nbusy.app;
+package com.nbusy.app.data;
+
+import com.nbusy.app.activities.ChatDetailFragment;
 
 import java.util.ArrayList;
 import java.util.Date;
