+++ app/src/main/java/com/nbusy/app/data/Chats.java
@@ -1,4 +1,4 @@
-package com.nbusy.app;
+package com.nbusy.app.data;
 
 import java.util.ArrayList;
 import java.util.Arrays;
