+++ app/src/main/java/com/nbusy/app/worker/UiThreadExecutor.java
@@ -1,4 +1,4 @@
-package com.nbusy.app;
+package com.nbusy.app.worker;
 
 import android.os.Handler;
 import android.os.Looper;
