+++ app/src/main/java/com/nbusy/app/worker/WorkerService.java
@@ -1,10 +1,12 @@
-package com.nbusy.app;
+package com.nbusy.app.worker;
 
 import android.app.Service;
 import android.content.Intent;
 import android.os.Binder;
 import android.os.IBinder;
 import android.util.Log;
+
+import com.nbusy.app.services.DeviceBootBroadcastReceiver;
 
 import java.util.Objects;
