+++ app/src/main/res/layout/activity_chat_detail.xml
@@ -1,4 +1,4 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:id="@+id/chat_detail_container"
     android:layout_width="match_parent" android:layout_height="match_parent"
-    tools:context=".ChatDetailActivity" tools:ignore="MergeRootFrame" />
+    tools:context=".activities.ChatDetailActivity" tools:ignore="MergeRootFrame" />
