+++ app/src/main/res/layout/activity_chat_list.xml
@@ -1,6 +1,6 @@
 <fragment xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:id="@+id/chat_list"
-    android:name="com.nbusy.app.ChatListFragment" android:layout_width="match_parent"
+    android:name="com.nbusy.app.activities.ChatListFragment" android:layout_width="match_parent"
     android:layout_height="match_parent" android:layout_marginLeft="16dp"
-    android:layout_marginRight="16dp" tools:context=".ChatListActivity"
+    android:layout_marginRight="16dp" tools:context=".activities.ChatListActivity"
     tools:layout="@android:layout/list_content" />
